+++ src/interpreters/Rust_original.rs
@@ -110,7 +110,7 @@ impl Interpreter for Rust_original {
 
     fn add_boilerplate(&mut self) -> Result<(), SniprunError> {
         if !Rust_original::contains_main("fn main", &self.code, "//") {
-            self.code = String::from("fn main() {") + &self.code + "}";
+            self.code = String::from("fn main() {") + &self.code + "\n}";
         }
         Ok(())
     }