+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v1.3.7
+- Fix issue in multiple languages when last line is a comment
+- Install instructions for lazy.nvim/Lunarvim
+- MSRV increase to 1.65
+- Bash interpreter fix + REPL feature removal
+- API's run_string function can display virtual text at current cursor position
+- HTTP support (using ureq)
+
 ## v1.3.6
 - Ruby documentation
 - Fix deprecating checkhealth